Add Rake to the mix and flesh out Cucumber tests
In order for Cucumber tests to be run correctly, a web server needs to
have the site running. So, I added rake tasks to the mix to handle :test

While the running of the Cucumber tests is good to go, there's still
a lingering issue that needs to be fixed - the kill command causes a
non-clean exit condition, and rake is aborted. Need to provide a way to
do this operation more cleanly. Suggestion was made by @fivetanley to
look into popen3 - I may look into this as a solution in the future.