Splitting out Rake tests and Travis script run info
Rake task :test has been split out into :test and :TravisTest.
The only difference between the two tasks is that :TravisTest
will not look for 'middleman' PIDs and do kill. This is done so that
Rake will not exit 0 and abort (PID will be killed when the Travis
spot instance is spun down following a CI run.)